requirements.dev.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cryptography==46.0.3
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.127.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
Flask==3.1.2
flask-cors==6.0.2
flask-talisman==1.1.0
fsspec==2025.12.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
markdown2==2.5.4
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
packaging==25.0
pdfkit==1.0.0
pluggy==1.6.0
postgrest==2.27.0
propcache==0.4.1
pycparser==2.23
pydantic==2.11.7
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.33.2
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.1
pyroaring==1.0.3
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
realtime==2.27.0
requests==2.32.5
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.47.2
storage3==2.27.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
tenacity==9.1.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
